WIP: Python example to be converted to notebooks
Examples are small python scripts checking that python-wrapped applications
can be used. They have to be converted to python notebooks to fit in the
TubeTK/Example dir.

To try them, scripts have to be run within an ITK virtual environment or by
making ITK-Wrapping available in the system path.
Example data can be found here :
https://drive.google.com/open?id=0B4vp_yoHdlN9WWQ4THByWWV4cWc